+++ web/app/services.php
@@ -51,6 +51,7 @@ $app['preferences.repository'] = $app->share(function($app) {
         'weekstart' => $app['defaults.weekstart'],
         'timezone' => $app['defaults.timezone'],
         'show_week_nb' => $app['defaults.show_week_nb'],
+        'show_now_indicator' => $app['defaults.show_now_indicator'],
     ]);
 
     return $repository;